module/vmanager/requirements.txt
# rest framework
Django==1.9.6
djangorestframework==3.3.3

# for call rest api
httplib2==0.9.2

# for unit test
coverage==4.2
mock==2.0.0
unittest_xml_reporting==1.12.0

# mysql connector
mysql-connector-python==8.0.15

# yaml processor
ruamel.yaml==0.15.85

# timer task
schedule==0.6.0

# http request
requests==2.21.0
